change coordinate type to double to get more precision

+++ src/Point3D.h
@@ -7,15 +7,15 @@
 
 class Point3D {
 public:
-    float x, y, z;
+    double x, y, z;
 
-    Point3D(float x = 0, float y = 0, float z = 0) : x(x), y(y), z(z) {}
+    Point3D(double x = 0, double y = 0, double z = 0) : x(x), y(y), z(z) {}
 
     Point3D operator-(const Point3D& other) const {
         return Point3D(x - other.x, y - other.y, z - other.z);
     }
 
-    float distanceTo(const Point3D& other) const {
+    double distanceTo(const Point3D& other) const {
         return std::sqrt((x - other.x) * (x - other.x) + 
                          (y - other.y) * (y - other.y) + 
                          (z - other.z) * (z - other.z));
